+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Changed
 
+- Pad based on textures and not channel props.
+
 ## 0.2.1
 
 ### Added